requirements.txt
django
mysqlclient
python-dotenv
djangorestframework
drf-yasg
djangorestframework-simplejwt